refacctor(ui): make icons scalable and reusable
